refactor : form modal 검증 함수를 Utils로 이동
- category, account form modal 검증 함수를 utils 폴더로 이동
